[#178] refactor: types 폴더 도메인 타입 이동 (Phase 11)
- types/record.ts → domains/record/types + domains/place/types
- types/auth.ts → domains/auth/types + domains/profile/types
- types/jjikboul.ts → domains/jjikboul/types
- types/common.ts 정리 (ClimbingPlace 이동, Pagination 유지)
- types/appScheme.ts 유지 (범용 타입)
- 모든 import 경로 수정
- 빌드 테스트 통과
