+++ src/constants/level.ts
@@ -1,4 +1,4 @@
-import { Level } from "@/types/record";
+import { Level } from '@/domains/place/types/entity';
 
 export const LEVELS: Level[] = [
   {
