+++ src/domains/place/api/transform.ts
@@ -1,5 +1,5 @@
 import type { PlaceResponse, LevelResponse } from '@/domains/place/types/response';
-import type { Place, Level } from '@/domains/place/types/entity';
+import type { Place, Level, LevelColor } from '@/domains/place/types/entity';
 
 export const transformPlaceResponseToEntity = (response: PlaceResponse): Place => ({
   id: response.gymId,
@@ -10,6 +10,6 @@ export const transformPlaceResponseToEntity = (response: PlaceResponse): Place =
 export const transformLevelResponseToEntity = (response: LevelResponse): Level => ({
   levelId: response.levelId,
   colorNameKo: response.colorNameKo ?? '',
-  colorNameEn: response.colorNameEn ?? '',
+  colorNameEn: (response.colorNameEn ?? 'grey') as LevelColor,
   colorCode: response.colorCode ?? '',
 });
