+++ src/domains/profile/types/entity.ts
@@ -1,5 +1,5 @@
-import type { OAuthProvider } from '@/domains/auth/types/entity';
-import type { RecordMetadata } from '@/types/record';
+import type { OAuthProvider, MemberInfo } from '@/domains/auth/types/entity';
+import type { RecordMetadata } from '@/domains/record/types/entity';
 
 export interface Profile {
   memberId: number;
@@ -8,6 +8,11 @@ export interface Profile {
   profileUrl: string;
   email: string;
 }
+
+// types/auth.ts에서 이동
+export type MyProfileApiResponse = Required<
+  Pick<MemberInfo, 'memberId' | 'nickname' | 'providerType' | 'profileUrl' | 'email'>
+>;
 
 export interface EditNicknameRequest {
   memberId: number;
