+++ src/domains/profile/types/response.ts
@@ -1,5 +1,5 @@
 import type { OAuthProvider } from '@/domains/auth/types/entity';
-import type { RecordMetadata } from '@/types/record';
+import type { RecordMetadata } from '@/domains/record/types/entity';
 import type { Pagination } from '@/types/common';
 
 export interface MyProfileResponse {
