+++ src/domains/record/components/Caution.tsx
