+++ src/domains/record/components/ClearButton.tsx
