+++ src/domains/record/components/RecordItem.tsx
@@ -1,10 +1,10 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import type { RecordMetadata } from "@/types/record";
+import type { RecordMetadata } from '@/domains/record/types/entity';
 import { fromNowFormat } from "@/utils/common";
 
-import LevelIcon from "@/components/common/LevelIcon";
+import LevelIcon from "@/domains/place/components/LevelIcon";
 import Avatar from "@/components/common/Avatar";
 
 export const RecordItemContainer = ({
