+++ src/domains/record/components/commonText.tsx
